doc_assistant_full/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
langchain==0.2.12
langchain-openai==0.1.22
faiss-cpu==1.8.0
pypdf==4.3.1
python-dotenv==1.0.1
